Add error checking to api

+++ src/types.rs
@@ -16,24 +16,24 @@ impl Stock {
 
 #[derive(Debug)]
 pub struct Order {
-    pub price: u64,
-    pub qty: u64
+    pub price: i32,
+    pub qty: i32
 }
 
 impl Order {
-    pub fn new(price: u64, qty: u64) -> Order {
+    pub fn new(price: i32, qty: i32) -> Order {
         Order{ price: price, qty: qty }
     }
 }
 
 #[derive(Debug)]
 pub struct Fill {
-    price: u64,
-    qty: u64
+    price: i32,
+    qty: i32
 }
 
 impl Fill {
-    pub fn new(price: u64, qty: u64) -> Fill {
+    pub fn new(price: i32, qty: i32) -> Fill {
         Fill{ price: price, qty: qty }
     }
 }
@@ -51,8 +51,7 @@ impl OrderBook {
         OrderBook{ venue: venue, symbol: symbol, bids: bids, asks: asks }
     }
 }
-
-#[allow(dead_code)] #[derive(Debug, Clone, RustcEncodable)]
+ #[derive(Debug, Clone, RustcEncodable)]
 pub enum Direction {
     Buy,
     Sell
@@ -67,8 +66,7 @@ impl Direction {
         }
     }
 }
-
-#[allow(dead_code)]  #[derive(Debug, Clone, RustcEncodable)]
+  #[derive(Debug, Clone, RustcEncodable)]
 pub enum OrderType {
     Limit,
     Market,
@@ -95,15 +93,15 @@ pub struct StockOrder {
     pub account: String,
     pub venue: String,
     pub symbol: String,
-    pub price: u64,
-    pub qty: u64,
+    pub price: i32,
+    pub qty: i32,
     pub direction: Direction,
     pub order_type: OrderType
 }
 
 impl StockOrder {
-    pub fn new(account: String, venue: String, symbol: String, price: u64,
-               qty: u64, direction: Direction, order_type: OrderType) -> StockOrder {
+    pub fn new(account: String, venue: String, symbol: String, price: i32,
+               qty: i32, direction: Direction, order_type: OrderType) -> StockOrder {
         StockOrder{ account: account, venue: venue, symbol: symbol, price: price,
                     qty: qty, direction: direction, order_type: order_type }
     }
@@ -125,20 +123,20 @@ impl Encodable for StockOrder {
 
 #[derive(Debug)]
 pub struct OrderStatus {
-    pub id: u64,
+    pub id: i32,
     pub venue: String,
     pub symbol: String,
     pub direction: Direction,
-    pub original_qty: u64,
-    pub total_filled: u64,
-    pub price: u64,
+    pub original_qty: i32,
+    pub total_filled: i32,
+    pub price: i32,
     pub order_type: OrderType,
     pub fills: Vec<Fill>
 }
 
 impl OrderStatus {
-    pub fn new(id: u64, venue: String, symbol: String, direction: Direction, original_qty: u64,
-               total_filled: u64, price: u64, order_type: OrderType, fills: Vec<Fill>) -> OrderStatus {
+    pub fn new(id: i32, venue: String, symbol: String, direction: Direction, original_qty: i32,
+               total_filled: i32, price: i32, order_type: OrderType, fills: Vec<Fill>) -> OrderStatus {
         OrderStatus{ id: id, venue: venue, symbol: symbol, direction: direction,
                      original_qty: original_qty, total_filled: total_filled, price: price, order_type: order_type,
                      fills: fills }
@@ -149,19 +147,19 @@ impl OrderStatus {
 pub struct StockQuote {
     pub venue: String,
     pub symbol: String,
-    pub bid: u64,       // best price currently bid for the stock
-    pub ask: u64,       // best price currently offered for the stock
-    pub bid_size: u64,  // aggregate size of all orders at the best bid
-    pub ask_size: u64,  // aggregate size of all orders at the best ask
-    pub bid_depth: u64, // aggregate size of *all bids*
-    pub ask_depth: u64, // aggregate size of *all asks*
-    pub last: u64,      // price of last trade
-    pub last_size: u64  // quantity of last trade
+    pub bid: i32,       // best price currently bid for the stock
+    pub ask: i32,       // best price currently offered for the stock
+    pub bid_size: i32,  // aggregate size of all orders at the best bid
+    pub ask_size: i32,  // aggregate size of all orders at the best ask
+    pub bid_depth: i32, // aggregate size of *all bids*
+    pub ask_depth: i32, // aggregate size of *all asks*
+    pub last: i32,      // price of last trade
+    pub last_size: i32  // quantity of last trade
 }
 
 impl StockQuote {
-    pub fn new(venue: String, symbol: String, bid: u64, ask: u64, bid_size: u64, ask_size: u64,
-               bid_depth: u64, ask_depth: u64, last: u64, last_size: u64) -> StockQuote {
+    pub fn new(venue: String, symbol: String, bid: i32, ask: i32, bid_size: i32, ask_size: i32,
+               bid_depth: i32, ask_depth: i32, last: i32, last_size: i32) -> StockQuote {
         StockQuote{ venue: venue, symbol: symbol, bid: bid, ask: ask,
                     bid_size: bid_size, ask_size: ask_size, bid_depth: bid_depth,
                     ask_depth: ask_depth, last: last, last_size: last_size }
